Issue-3: Moved ExpandableLayout example to separate module.
