+++ app/src/main/res/values/strings.xml
@@ -19,8 +19,6 @@
     <string name="duo.navigation.drawer">DuoNavigationDrawer</string>
     <string name="duo.navigation.menu.item">Menu item</string>
 
-    <string name="expandable.linear.layout">Expandable LinearLayout</string>
-
     <string name="vega.layout.manager">Vega Layout Manager</string>
 
     <string name="fancy.toast">Fancy Toast</string>
