+++ expandablelayout/src/main/AndroidManifest.xml
@@ -1,0 +1,10 @@
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.kirchhoff.expandablelayout">
+
+    <application>
+        <activity android:name=".ExpandableLayoutActivity"
+            android:label="@string/expandable_layout_title"/>
+    </application>
+
+</manifest>
+
