+++ expandablelayout/src/main/res/layout/a_expandable_linear_layout.xml
