+++ expandablelayout/src/main/res/layout/r_expandable_linear_layout_child.xml
